requirements.txt
asgiref==3.8.1
boto3==1.38.8
botocore==1.38.8
Django==5.2
django-storages==1.14.6
jmespath==1.0.1
pillow==11.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
gunicorn==21.2.0 
